+++ Ogma3/Data/DTOs/ShelfFromApiDTO.cs
@@ -21,7 +21,7 @@ namespace Ogma3.Data.DTOs
         public string? Icon { get; set; }
         public bool? DoesContainBook { get; set; }
 
-        public static ShelfFromApiDTO FromShelf(Shelf shelf, int? bookId = null) =>
+        public static ShelfFromApiDTO FromShelf(Shelf shelf, long? bookId = null) =>
             new ShelfFromApiDTO
             {
                 Id = shelf.Id,
